backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy>=2.0.35
beautifulsoup4==4.12.2
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
anthropic>=0.40.0
feedparser==6.0.10
python-dateutil==2.8.2
selenium==4.16.0
webdriver-manager==4.0.1
